Update: migrate to decimal type in TariffEntity class
Refs: (#2)

+++ database/07-accounting.sql
@@ -44,7 +44,7 @@ create table if not exists Accounting.Tariff
     schoolYear varchar(20) not null,
     educationalLevel smallint not null,
     concept varchar(200) not null,
-    amount float not null,
+    amount decimal(18, 2) not null,
     dueDate date,
     late boolean,
     typeId int not null,
